test was not working on windows

+++ jstests/shellkillop.js
@@ -1,33 +1,48 @@
 baseName = "jstests_shellkillop";
 
-db[ baseName ].drop();
-
-print("shellkillop.js insert data");
-for (i = 0; i < 100000; ++i) {
-    db[ baseName ].save( {i:1} );
+if (_isWindows()) {
+    print("shellkillop.js not testing on windows, as functionality is missing there");
+    print("shellkillop.js see http://jira.mongodb.org/browse/SERVER-1451");
 }
-assert.eq( 100000, db[ baseName ].count() );
+else {
 
-var evalStr = "print('SKO subtask started'); db." + baseName + ".update( {}, {$set:{i:'abcdefghijkl'}}, false, true ); db." + baseName + ".count();";
-print("shellkillop.js evalStr:" + evalStr);
-spawn = startMongoProgramNoConnect("mongo", "--autokillop", "--port", myPort(), "--eval", evalStr);
+    db[baseName].drop();
 
-sleep(100);
-
-stopMongoProgramByPid(spawn);
-
-sleep(200);
-var inprog = db.currentOp().inprog;
-if (inprog.count) {
-    sleep(2000); // wait more if need be
-    inprog = db.currentOp().inprog;
-}
-for( i in inprog ) {
-    if (inprog[i].ns == "test." + baseName) {
-        print("shellkillop.js FAIL op is still running:");
-        printjson(inprog);
-        assert(false, "still running op");
+    print("shellkillop.js insert data");
+    for (i = 0; i < 100000; ++i) {
+        db[baseName].insert({ i: 1 });
     }
-}
+    assert.eq(100000, db[baseName].count());
 
-print("shellkillop.js SUCCESS");
+    // mongo --autokillop suppressed the ctrl-c "do you want to kill current operation" message
+    // it's just for testing purposes and thus not in the shell help
+    var evalStr = "print('SKO subtask started'); db." + baseName + ".update( {}, {$set:{i:'abcdefghijkl'}}, false, true ); db." + baseName + ".count();";
+    print("shellkillop.js evalStr:" + evalStr);
+    spawn = startMongoProgramNoConnect("mongo", "--autokillop", "--port", myPort(), "--eval", evalStr);
+
+    sleep(100);
+    assert(db[baseName].find({ i: 'abcdefghijkl' }).count() < 100000, "update ran too fast, test won't be valid");
+
+    stopMongoProgramByPid(spawn);
+
+    sleep(200);
+
+    print("count abcdefghijkl:" + db[baseName].find({ i: 'abcdefghijkl' }).count());
+
+    assert.soon(
+    function f() {
+        var inprog = db.currentOp().inprog;
+        for (i in inprog) {
+            if (inprog[i].ns == "test." + baseName)
+                print("shellkillop.js op is still running, waiting");
+            //printjson(inprog);
+            return false;
+        }
+        return true;
+    },
+    "shellkillop.js FAIL op is still running"
+    );
+
+    print("shellkillop.js SUCCESS");
+
+}
